+++ src/test/java/ru/netology/stats/StatsServiceTest.java
@@ -11,7 +11,7 @@ public class StatsServiceTest {
 
         StatsService service = new StatsService();
         int expected = 180;
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         long actuall = service.summOfAllSales(sales);
         Assertions.assertEquals(expected, actuall);
 
@@ -22,7 +22,7 @@ public class StatsServiceTest {
 
         StatsService service = new StatsService();
         int expected = 15;
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         long actuall = service.averageSumm(sales);
         Assertions.assertEquals(expected, actuall);
     }
@@ -31,7 +31,7 @@ public class StatsServiceTest {
     public void maxSalestest() {
         StatsService service = new StatsService();
         int expected = 8;
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int actuall = service.maxSales(sales);
         Assertions.assertEquals(expected, actuall);
     }
@@ -40,7 +40,7 @@ public class StatsServiceTest {
     public void minSalestest() {
         StatsService service = new StatsService();
         int expected = 9;
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int actuall = service.minSales(sales);
         Assertions.assertEquals(expected, actuall);
     }
@@ -49,7 +49,7 @@ public class StatsServiceTest {
     public void amountMonthsBelowtest() {
         StatsService service = new StatsService();
         int expected = 5;
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int actuall = service.amountMonthsBelow(sales);
         Assertions.assertEquals(expected, actuall);
     }
@@ -58,7 +58,7 @@ public class StatsServiceTest {
     public void amountMonthsUndertest() {
         StatsService service = new StatsService();
         int expected = 5;
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int actuall = service.amountMonthsUnder(sales);
         Assertions.assertEquals(expected, actuall);
     }
